refactor: further tweaks to make things work with jest
