+++ src/context-helpers/context-manager.ts
@@ -1,20 +1,17 @@
-import { AsyncLocalStorage } from 'node:async_hooks'
 import type { TestExecutionContext } from '../test-execution-context'
 
 export class ContextManager {
-  private static asyncStore = new AsyncLocalStorage<TestExecutionContext>()
+  private static _instance: TestExecutionContext | undefined
 
   static set instance(ctx: TestExecutionContext) {
-    const instance = this.asyncStore.getStore()
-    if (instance !== undefined) throw new Error('Execution context has already been set')
-    this.asyncStore.enterWith(ctx)
+    if (this._instance !== undefined) throw new Error('Execution context has already been set')
+    this._instance = ctx
   }
   static get instance() {
-    const instance = this.asyncStore.getStore()
-    if (instance === undefined) throw new Error('No execution context has been set')
-    return instance
+    if (this._instance === undefined) throw new Error('No execution context has been set')
+    return this._instance
   }
   static reset() {
-    this.asyncStore.disable()
+    this._instance = undefined
   }
 }
