+++ src/internal/arc4.ts
@@ -1,0 +1,2 @@
+export * from '@algorandfoundation/algorand-typescript/arc4'
+export { abimethod, Contract, methodSelector } from '../impl/contract'
